Move TS packaging to top level
