+++ .gitignore
@@ -36,3 +36,19 @@ smithy-typescript-integ-tests/yarn.lock
 
 # Issue https://github.com/awslabs/smithy-typescript/issues/425
 smithy-typescript-codegen/bin/
+
+**/node_modules/
+**/*.tsbuildinfo
+**/*.d.ts
+**/yarn-error.log
+coverage
+**/.turbo
+dist-*
+*.tsbuildinfo
+.pnp.*
+.yarn/*
+!.yarn/patches
+!.yarn/plugins
+!.yarn/releases
+!.yarn/sdks
+!.yarn/versions
