+++ package.json
@@ -22,18 +22,17 @@
     "rimraf": "^3.0.2"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.4",
-    "jest": "^26.1.0",
+    "@types/jest": "28.1.3",
+    "jest": "28.1.1",
+    "jest-environment-jsdom": "28.1.1",
     "prettier": "2.2.1",
-    "ts-jest": "^26.4.1",
+    "ts-jest": "28.0.5",
     "turbo": "^1.1.2",
-    "typescript": "~4.2.4",
+    "typescript": "~4.9.5",
     "yarn": "1.22.10"
   },
-  "workspaces": {
-    "packages": [
-      "*"
-    ]
-  },
+  "workspaces": [
+    "smithy-typescript-ssdk-libs/*"
+  ],
   "packageManager": "yarn@3.4.1"
 }
