--- smithy-typescript-ssdk-libs/.gitignore
@@ -1,13 +1,0 @@
-**/node_modules/
-**/*.tsbuildinfo
-**/*.d.ts
-**/yarn-error.log
-coverage
-**/.turbo
-.pnp.*
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/sdks
-!.yarn/versions
