--- smithy-typescript-ssdk-libs/jest.config.base.js
@@ -1,13 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  transform: {
-    "^.+\\.(ts|tsx)$": "ts-jest",
-  },
-  globals: {
-    "ts-jest": {
-      tsconfig: "tsconfig.json",
-    },
-  },
-  testMatch: ["**/*.spec.ts", "!**/*.integ.spec.ts"],
-};
