+++ smithy-typescript-ssdk-libs/server-apigateway/jest.config.js
@@ -1,4 +1,4 @@
-const base = require("../jest.config.base.js");
+const base = require("../../jest.config.base.js");
 
 module.exports = {
   preset: "ts-jest",
