+++ smithy-typescript-ssdk-libs/server-common/package.json
@@ -32,7 +32,6 @@
     "tslib": "^1.8.0"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.22",
     "@types/node": "^14.14.37",
     "@typescript-eslint/eslint-plugin": "4.30.0",
     "@typescript-eslint/parser": "4.30.0",
@@ -42,10 +41,8 @@
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-prettier": "3.4.1",
     "eslint-plugin-simple-import-sort": "7.0.0",
-    "jest": "^26.6.3",
     "rimraf": "^3.0.2",
-    "ts-jest": "^26.5.2",
-    "typescript": "^4.2.2"
+    "typescript": "^4.9.5"
   },
   "files": [
     "dist/cjs/**/*.js",
