+++ smithy-typescript-ssdk-libs/server-node/jest.config.js
@@ -1,4 +1,4 @@
-const base = require("../jest.config.base.js");
+const base = require("../../jest.config.base.js");
 
 module.exports = {
   preset: "ts-jest",
