+++ smithy-typescript-ssdk-libs/server-node/package.json
@@ -31,7 +31,6 @@
     "tslib": "^1.8.0"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.4",
     "@types/node": "^14.14.37",
     "concurrently": "^7.6.0",
     "downlevel-dts": "^0.7.0",
@@ -39,9 +38,8 @@
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-prettier": "3.4.1",
     "eslint-plugin-simple-import-sort": "7.0.0",
-    "jest": "^26.1.0",
     "rimraf": "^3.0.2",
-    "typescript": "~4.2.2"
+    "typescript": "~4.9.5"
   },
   "files": [
     "dist/cjs/**/*.js",
