+++ smithy-typescript-ssdk-libs/server-node/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "stripInternal": true,
     "removeComments": true,
