+++ tsconfig.cjs.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "importHelpers": true,
+    "module": "commonjs",
+    "noEmitHelpers": true,
+    "target": "ES2018",
+    "strict": true,
+    "strictNullChecks": true
+  }
+}
