+++ tsconfig.es.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "importHelpers": true,
+    "module": "esnext",
+    "noEmitHelpers": true,
+    "target": "ES2020",
+    "strict": true,
+    "strictNullChecks": true
+  }
+}
