+++ tsconfig.json
@@ -15,6 +15,8 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "target": "ES2018",
+    "useUnknownInCatchVariables": false,
+    "strictNullChecks": false,
     "lib": ["es2018", "dom"],
     "baseUrl": ".",
     "strict": true,
@@ -27,5 +29,6 @@
       "@aws-smithy/server-node": ["server-node/src"]
     }
   },
-  "exclude": ["node_modules/"]
+  "include": ["packages/", "smithy-typescript-ssdk-libs/"],
+  "exclude": ["node_modules/", "**/*.spec.ts"]
 }
